Refactor of Controller. Now OperatorService + Deletion & Reconciliation controllers.
