+++ docker/cassandra/cassandra.sh
@@ -1,29 +1,22 @@
-#!/bin/bash -ue
+#!/bin/bash -xue
 
-echoerr() { echo "$@" 1>&2; }
+shopt -s nullglob
 
 . /usr/share/cassandra/cassandra.in.sh
 
 JVM_OPTS=${JVM_OPTS:=}
 
-## sanity checks
-#for conf in "${CASSANDRA_CONF}/cassandra.yaml" "${CASSANDRA_CONF}/cassandra-env.sh"; do
-#    if [ ! -f "${conf}" ]; then
-#        echoerr "${conf}: File not found. Required to start Cassandra.";
-#        exit 1;
-#    fi
-#done
-
-
 JVM_OPTS="${JVM_OPTS} -Dcassandra.config.loader=com.instaclustr.cassandra.k8s.ConcatenatedYamlConfigurationLoader"
-JVM_OPTS="${JVM_OPTS} -Dcassandra.config=/etc/cassandra.yaml.d/operator:/etc/cassandra.yaml.d/user"
-JVM_OPTS="${JVM_OPTS} -Dcassandra.storagedir=/var/lib/cassandra" # set via YAML
+JVM_OPTS="${JVM_OPTS} -Dcassandra.config=/etc/cassandra/cassandra.yaml:/etc/cassandra/cassandra.yaml.d"
 
 # provides hints to the JIT compiler
-JVM_OPTS="${JVM_OPTS} -XX:CompileCommandFile=/usr/share/cassandra/hotspot_compiler"
+JVM_OPTS="${JVM_OPTS} -XX:CompileCommandFile=${CASSANDRA_HOME}/hotspot_compiler"
 
-# add the jamm javaagent
+# add the jamm agent
 JVM_OPTS="${JVM_OPTS} -javaagent:${CASSANDRA_HOME}/agents/jamm-0.3.0.jar"
+
+## Prometheus exporter agent
+#JVM_OPTS="${JVM_OPTS} -javaagent:${CASSANDRA_HOME}/agents/cassandra-exporter-agent-0.9.0.jar"
 
 # sigar support
 JVM_OPTS="${JVM_OPTS} -Djava.library.path=${CASSANDRA_HOME}/lib/sigar-bin"
@@ -32,14 +25,20 @@ JVM_OPTS="${JVM_OPTS} -Djava.library.path=${CASSANDRA_HOME}/lib/sigar-bin"
 JVM_OPTS="${JVM_OPTS} -XX:HeapDumpPath=/var/tmp/cassandra-`date +%s`-pid$$.hprof"
 
 
-# Read additional JVM options from jvm.options file
-JVM_OPTS="${JVM_OPTS} "$((sed -ne "/^-/p" | tr '\n' ' ') < ${CASSANDRA_CONF}/jvm.options)
+# read additional JVM options from jvm.options files
+for options_file in "${CASSANDRA_CONF}/jvm.options" "${CASSANDRA_CONF}/jvm.options.d"/*
+do
+    JVM_OPTS="${JVM_OPTS} "$((sed -ne "/^-/p" | tr '\n' ' ') < "${options_file}")
+done
+
+# source additional environment settings
+for env_file in "${CASSANDRA_CONF}/cassandra-env.sh" "${CASSANDRA_CONF}/cassandra-env.sh.d"/*
+do
+    . "${env_file}"
+done
 
 
-. ${CASSANDRA_CONF}/cassandra-env.sh
-
-
-exec -a cassandra /usr/bin/java \
+exec /usr/bin/java \
     -cp "${CASSANDRA_CLASSPATH}" \
     ${JVM_OPTS} \
     -Dcassandra-foreground=yes \
