+++ docker/cassandra/default-config/cassandra-env.sh
@@ -1,0 +1,1 @@
+# empty by default
