+++ docker/cassandra/default-config/jvm.options
@@ -1,0 +1,42 @@
+-Dcassandra.jmx.remote.port=7199
+-Dcom.sun.management.jmxremote.authenticate=false
+
+-Dcassandra.libjemalloc=/usr/lib64/libjemalloc.so.1
+
+-XX:+UseThreadPriorities
+-XX:ThreadPriorityPolicy=42
+
+-Xss256k
+
+-XX:StringTableSize=1000003
+
+-XX:+AlwaysPreTouch
+-XX:-UseBiasedLocking
+
+-XX:+UseTLAB
+-XX:+ResizeTLAB
+
+-XX:+PerfDisableSharedMem
+
+-Djava.net.preferIPv4Stack=true
+
+
+# GC logging
+-XX:+PrintGCDetails
+-XX:+PrintGCDateStamps
+-XX:+PrintHeapAtGC
+-XX:+PrintTenuringDistribution
+-XX:+PrintGCApplicationStoppedTime
+-XX:+PrintPromotionFailure
+-Xloggc:/var/lib/cassandra/logs/gc.log
+-XX:+UseGCLogFileRotation
+-XX:NumberOfGCLogFiles=10
+-XX:GCLogFileSize=10M
+
+-XX:+PrintFlagsFinal
+-XX:+HeapDumpOnOutOfMemoryError
+
+-XX:MaxDirectMemorySize=18G
+
+
+# GC settings are not defined here -- they are chosen by the operator at runtime
