+++ java/model/src/main/resources/schema/CassandraDataCenter.json
@@ -56,8 +56,9 @@
           "type" : "string",
           "description" : "Name of the CassandraBackup to restore from"
         },
-        "userConfigMap": {
-          "type" : "string",
+        "userConfigMapVolumeSource": {
+          "type" : "object",
+          "javaType": "io.kubernetes.client.models.V1ConfigMapVolumeSource",
           "description" : "Name of an optional config map that contains cassandra configuration in the form of yaml fragments"
         }
       }
