+++ java/operator/src/main/java/com/instaclustr/cassandra/operator/OperatorModule.java
@@ -2,11 +2,13 @@ package com.instaclustr.cassandra.operator;
 
 import com.google.common.util.concurrent.Service;
 import com.google.inject.AbstractModule;
+import com.google.inject.assistedinject.FactoryModuleBuilder;
 import com.google.inject.multibindings.Multibinder;
+import com.instaclustr.cassandra.operator.controller.DataCenterControllerFactory;
 import com.instaclustr.cassandra.operator.modules.*;
 import com.instaclustr.cassandra.operator.service.BackupControllerService;
 import com.instaclustr.cassandra.operator.service.CassandraHealthCheckService;
-import com.instaclustr.cassandra.operator.service.ControllerService;
+import com.instaclustr.cassandra.operator.service.OperatorService;
 import com.instaclustr.cassandra.operator.service.GarbageCollectorService;
 
 public class OperatorModule extends AbstractModule {
@@ -15,10 +17,12 @@ public class OperatorModule extends AbstractModule {
     protected void configure() {
         final Multibinder<Service> serviceMultibinder = Multibinder.newSetBinder(binder(), Service.class);
 
-        serviceMultibinder.addBinding().to(ControllerService.class);
+        serviceMultibinder.addBinding().to(OperatorService.class);
         serviceMultibinder.addBinding().to(GarbageCollectorService.class);
         serviceMultibinder.addBinding().to(CassandraHealthCheckService.class);
         serviceMultibinder.addBinding().to(BackupControllerService.class);
+
+        install(new FactoryModuleBuilder().build(DataCenterControllerFactory.class));
 
         install(new ClusterWatchModule());
         install(new DataCenterWatchModule());
