--- java/operator/src/main/resources/com/instaclustr/cassandra/config/cassandra-env.sh
@@ -1,1 +1,0 @@
-# empty by default. Append to JVM_OPTS to add additional JVM options
