--- java/operator/src/main/resources/com/instaclustr/cassandra/config/jvm.options
@@ -1,90 +1,0 @@
--Xms128M
--Xmx128M
-
--Dcassandra.jmx.remote.port=7199
--Dcom.sun.management.jmxremote.authenticate=false
-
--Dcassandra.consistent.rangemovement=false
-
--Dcassandra.libjemalloc=/usr/lib64/libjemalloc.so.1
-
--Dcassandra.metricsReporterConfigFile=metrics-reporter-config.yaml
-
--XX:+UseThreadPriorities
--XX:ThreadPriorityPolicy=42
-
--Xss256k
-
--XX:StringTableSize=1000003
-
--XX:+AlwaysPreTouch
--XX:-UseBiasedLocking
-
--XX:+UseTLAB
--XX:+ResizeTLAB
-
--XX:+PerfDisableSharedMem
-
--Djava.net.preferIPv4Stack=true
-
-# GC Settings
-
--XX:ParallelGCThreads=14
--XX:ConcGCThreads=14
-
-## G1 GC Settings
-#-XX:+UseG1GC
-#
-#-XX:G1RSetUpdatingPauseTimePercent=5
-#-XX:InitiatingHeapOccupancyPercent=70
-##-XX:MaxGCPauseMillis=500
-#
-#-XX:+UseCondCardMark
-
--XX:+UnlockDiagnosticVMOptions
--XX:+UseGCTaskAffinity
--XX:+BindGCTaskThreadsToCPUs
--XX:ParGCCardsPerStrideChunk=32768
--XX:-UseBiasedLocking
-
-## CMS GC settings
--XX:+CMSClassUnloadingEnabled
-
--XX:+UseParNewGC
--XX:+UseConcMarkSweepGC
--XX:+CMSParallelRemarkEnabled
--XX:SurvivorRatio=2
--XX:MaxTenuringThreshold=15
--XX:CMSInitiatingOccupancyFraction=75
--XX:+UseCMSInitiatingOccupancyOnly
--XX:+UseTLAB
--XX:+UseCondCardMark
--XX:+CMSParallelInitialMarkEnabled
--XX:+CMSEdenChunksRecordAlways
--XX:CMSWaitDuration=10000
-
--XX:+ParallelRefProcEnabled
-
-
-
-
-# GC logging
--XX:+PrintGCDetails
--XX:+PrintGCDateStamps
--XX:+PrintHeapAtGC
--XX:+PrintTenuringDistribution
--XX:+PrintGCApplicationStoppedTime
--XX:+PrintPromotionFailure
--Xloggc:/var/log/cassandra/gc.log
--XX:+UseGCLogFileRotation
--XX:NumberOfGCLogFiles=10
--XX:GCLogFileSize=10M
-
-
--XX:+PrintFlagsFinal
--XX:+HeapDumpOnOutOfMemoryError
-
-#-Djdk.nio.maxCachedBufferSize=262144
-
-
--XX:MaxDirectMemorySize=18G
